adjustments on code

+++ src/models/property.py
@@ -5,5 +5,5 @@ from src.models.player import Player
 class Property(BaseModel):
     id: int
     owner: Player = None
-    price: int = random.randint(20, 110)
-    rent_price: int = price/2
+    price: float = random.randint(20, 110)
+    rent_price: float = price/2
